+++ src/main/java/org/eclipse/keyple/distributed/AbstractRemoteReaderAdapter.java
@@ -26,6 +26,7 @@ abstract class AbstractRemoteReaderAdapter implements RemoteReaderSpi {
   private final int clientCoreApiLevel;
   private final String remoteReaderName;
   private final String localReaderName;
+  private final Boolean isContactless;
   private final String sessionId;
   private final String clientNodeId;
   private final AbstractNodeAdapter node;
@@ -37,6 +38,7 @@ abstract class AbstractRemoteReaderAdapter implements RemoteReaderSpi {
    * @param clientCoreApiLevel The API level of the client Core layer.
    * @param remoteReaderName The name of the remote reader.
    * @param localReaderName The name of the associated local reader.
+   * @param isContactless Is local reader contactless (null if unknown).
    * @param sessionId The associated session ID.
    * @param clientNodeId The associated client node ID.
    * @param node The associated node.
@@ -47,6 +49,7 @@ abstract class AbstractRemoteReaderAdapter implements RemoteReaderSpi {
       int clientCoreApiLevel,
       String remoteReaderName,
       String localReaderName,
+      Boolean isContactless,
       String sessionId,
       String clientNodeId,
       AbstractNodeAdapter node) {
@@ -54,6 +57,7 @@ abstract class AbstractRemoteReaderAdapter implements RemoteReaderSpi {
     this.clientCoreApiLevel = clientCoreApiLevel;
     this.remoteReaderName = remoteReaderName;
     this.localReaderName = localReaderName;
+    this.isContactless = isContactless;
     this.sessionId = sessionId;
     this.clientNodeId = clientNodeId;
     this.node = node;
@@ -123,6 +127,16 @@ abstract class AbstractRemoteReaderAdapter implements RemoteReaderSpi {
   @Override
   public final String getName() {
     return remoteReaderName;
+  }
+
+  /**
+   * {@inheritDoc}
+   *
+   * @since 2.5.0
+   */
+  @Override
+  public final Boolean isContactless() {
+    return isContactless;
   }
 
   /**