+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Upgraded
+- Keyple Distributed Network Lib `2.4.0` -> `2.5.0`
+- Keyple Distributed Remote API `3.0.1` -> `3.1.0`
 
 ## [2.4.0] - 2024-06-03
 ### Added